refactor: split diagnostic.py into data model and rendering modules
Separate the 870-line diagnostic.py into two focused files:
- diagnostic.py (~220 lines): Diagnostic type, Severity enum, dedup,
  exception_to_diagnostics, _CATEGORY_TITLES
- diagnostic_render.py (~650 lines): format_diagnostic and all
  _format_*/_render_* text rendering helpers

Split boundary: producers of Diagnostic stay with the type;
consumers (text rendering) move to diagnostic_render.py.

exception_to_diagnostics stays in diagnostic.py because it's a
factory that creates Diagnostic objects (not rendering), and is
used by execution/runner.py (not display-layer code).

Update 11 source files + 8 test files + 4 CLAUDE.md files.
Zero logic changes — code moved between files, nothing else.
