+++ src/pflow/cli/workflow_output.py
@@ -8,7 +8,8 @@ from typing import Any
 
 import click
 
-from pflow.core.diagnostic import Diagnostic, format_diagnostic
+from pflow.core.diagnostic import Diagnostic
+from pflow.core.diagnostic_render import format_diagnostic
 
 
 def safe_output(value: Any) -> bool:
